+++ desmume/src/gfx3d.cpp
@@ -1363,7 +1363,7 @@ void gfx3d_glLoadIdentity()
 
 BOOL gfx3d_glLoadMatrix4x4(s32 v)
 {
-	mtxCurrent[mode][ML4x4ind] = v;
+	mtxCurrent[mode][ML4x4ind] = (float)v;
 
 	++ML4x4ind;
 	if(ML4x4ind<16) return FALSE;
@@ -1380,7 +1380,7 @@ BOOL gfx3d_glLoadMatrix4x4(s32 v)
 
 BOOL gfx3d_glLoadMatrix4x3(s32 v)
 {
-	mtxCurrent[mode][ML4x3ind] = v;
+	mtxCurrent[mode][ML4x3ind] = (float)v;
 
 	ML4x3ind++;
 	if((ML4x3ind & 0x03) == 3) ML4x3ind++;
@@ -1402,7 +1402,7 @@ BOOL gfx3d_glLoadMatrix4x3(s32 v)
 
 BOOL gfx3d_glMultMatrix4x4(s32 v)
 {
-	mtxTemporal[MM4x4ind] = v;
+	mtxTemporal[MM4x4ind] = (float)v;
 
 	MM4x4ind++;
 	if(MM4x4ind<16) return FALSE;
@@ -1426,7 +1426,7 @@ BOOL gfx3d_glMultMatrix4x4(s32 v)
 
 BOOL gfx3d_glMultMatrix4x3(s32 v)
 {
-	mtxTemporal[MM4x3ind] = v;
+	mtxTemporal[MM4x3ind] = (float)v;
 
 	MM4x3ind++;
 	if((MM4x3ind & 0x03) == 3) MM4x3ind++;
@@ -1456,7 +1456,7 @@ BOOL gfx3d_glMultMatrix4x3(s32 v)
 
 BOOL gfx3d_glMultMatrix3x3(s32 v)
 {
-	mtxTemporal[MM3x3ind] = v;
+	mtxTemporal[MM3x3ind] = (float)v;
 
 
 	MM3x3ind++;
